requirements.txt
# ─── Core LLM & Agents ────────────────────────────────────────────────────────
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0

# ─── Vector Database ──────────────────────────────────────────────────────────
faiss-cpu>=1.7.4

# ─── Backend ──────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ─── PDF Processing ───────────────────────────────────────────────────────────
pypdf>=4.2.0

# ─── Embeddings ───────────────────────────────────────────────────────────────
tiktoken>=0.7.0

# ─── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.32.0
sseclient-py>=1.8.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
numpy>=1.26.0
httpx>=0.27.0
